requirements-eval.txt
# Eval framework additional dependencies (install alongside requirements.txt)
pyyaml>=6.0
